Use :restart instead of :start for procps service
It was noted in #31 that the defined sysctl settings were not being applied when
using the default recipe on CentOS 7. I noticted the same behavior in Suse 12,
and determined that restarting the service instead of just issuing a start
resolved the problem. I suspect this has something to do with how systemd is
handling the start command with regard to 'static' services.

In #15 there was question on whether a restart was safe. After running through
the tests, all of the platforms which converge successfully pass tests with
restart in place.