Remove mentions of attributes (#103)
* Remove mentions of attributes

* Rename test cookbook to test

* update ohai to 5+ to remove compat resource

* Add another test for removal

* Remove a template file that's for an unsupported platform

* Move the plugin up a directory

* Indent, and remove unused methods

* Turn template into file

* Move all helpers into helpers.rb

* fix travis test names

* - set_param now doesn't use attributes
- always ignore error when getting a key, that way the erorr is vomited
back into the Chef run if there is one.
- Add the ignore error property back in from the attributes
- Execute sysctl --system to temporarily set an attribute. between
reboots (makes sysctl re-evealuate all the sysctl.d properties)
- Stopped using |= because it wasn't doing what it should. So using nil?
instead (which technically is the same thing)

This mean we can set a sysctl param twice in one run. Then remove it
later on. And it be back to the original value.

* Fix reload resource for systemd

* Amazon is centos-6 with with less features...

Use -p because amazon-1 is super old

* Remove the file test if we're in docker as we aren't able to create that
file in Docker (works fine in dev though ;) )

* Fix that Travis error

* Remove tests the used mixlib shell out. Push this responsability to
integration tests.

Remove README information about attributes and and some more information
about custom resources

* Remove node.save.